Initial look-ahead.
This is an initial attempt on look-ahead similar to Marlins: allowing a small jerk to keep moving.

This version is INCOMPLETE: it only checks if the angles of 2 moves are similar, and if so, it removes the rampup and rampdown. This means that if you leave the feed rate alone between moves, it will give you full-speed transitions between moves in the same direction.

Todo:
* re-calculate the rampup and rampdown parameters to join 2 moves with different flow rates.
* implement the notion of jerk instead of similar angles: even if the angles are similar (within 30 degrees), at 4000mm/min you do not want to head off in a 30 degree direction without slowing down first.