Reverse pass added to lookahead.
This means that if the last move in the queue could not do a full stop when it ends, the crossing speeds between both moves is lowered and the second to last move should be modified to ramp up and down from its starting speed and ending speed.

Tested with:
- Multiple move lengths and mix of small and large moves
- Various feed rates

This still generates a load of debug output but its functional.