Fixes to lookahead
Fixed up and down ramps when lookahead is not used; it now works like it used to when lookahead it turned off.

Fixed missing limiter on speed in previous move; if the move was longer than needed to reach the target speed, it would keep accelerating. This resulted in weird results in other calculations.

Added various sanity checks and debug output. This will be removed in future commits; the sanity checks (sans loads of output) will most likely be optional to aid debugging issues.