Added missing speed limit to lookahead.
During the speed calculations in for the previous move, the crossing speed was not limited to the speed of the current move. This resulted in weird calculation errors in further logic.

Added defines for explicit lookahead debugging and verbose debugging. Most of the checks still need to be cleaned up and put between #ifdefs.

Todo: retain which axis was leading for a move and use the correct steps per m for that move (currently tested with X and Y with identical steps per m).