Switched ramp calculation to use X axis and implemented 1 dimensional jerk for extruder axis
The current linear acceleration logic uses the X axis for the acceleration parameters. As such, we can not calculate ramp lengths based on the leading axis: they are always based on the X axis.

TODO: Test if acceleration with X and E is 5 times faster than X alone - it should be because the current code ramps up every timer tick.

Implemented a 1D jerk calculation to determine the jerk on the extruder during move joining.